RebornUrSelf/requirements.txt
django>=5.2,<6.0
google-genai>=1.0
Pillow>=10.0
python-dotenv>=1.0
